Align server and firmware build versions
